+++ src/check.rs
@@ -425,18 +425,29 @@ impl<'a> VirtualEnv<'a> {
                     self.handle_sig(f.signature())?;
                 }
                 Try => {
-                    let f = self.pop_func()?;
-                    let handler = self.pop_func()?;
-                    let f_sig = f.signature();
-                    let target_handler_sig = Signature::new(f_sig.args + 1, f_sig.outputs);
-                    let handler_sig = handler.signature();
-                    if !handler_sig.is_subset_of(target_handler_sig) {
-                        return Err(format!(
-                            "try's functions have signatures {f_sig} and {handler_sig}, but \
-                            the error handler should take one more argument than the function."
-                        )
+                    let f_sig = self.pop_func()?.signature();
+                    let handler_sig = self.pop_func()?.signature();
+                    if handler_sig.args == 0 || handler_sig.args == 1 {
+                        0
+                    } else if handler_sig.args == f_sig.args + 1 {
+                        f_sig.args
+                    } else {
+                        return Err(if f_sig.args == 0 {
+                            format!(
+                                "handler function must take, 0 or 1 arguments, \
+                                but it takes {}.",
+                                handler_sig.args
+                            )
+                        } else {
+                            format!(
+                                "handler function must take, 0, 1, or {} arguments, \
+                                but it takes {}.",
+                                f_sig.args + 1,
+                                handler_sig.args
+                            )
+                        }
                         .into());
-                    }
+                    };
                     self.handle_sig(f_sig)?;
                 }
                 Fill => {