+++ src/primitive/mod.rs
@@ -596,22 +596,7 @@ impl Primitive {
                 env.call_with_this(f)?;
             }
             Primitive::Recur => env.recur()?,
-            Primitive::Try => {
-                let f = env.pop_function()?;
-                let handler = env.pop_function()?;
-                let f_args = f.signature().args;
-                let backup = env.clone_stack_top(f_args);
-                if let Err(e) = env.call_clean_stack(f) {
-                    env.rt
-                        .backend
-                        .save_error_color(e.message(), e.report().to_string());
-                    env.push(e.value());
-                    for val in backup {
-                        env.push(val);
-                    }
-                    env.call(handler)?;
-                }
-            }
+            Primitive::Try => algorithm::try_(env)?,
             Primitive::Assert => {
                 let msg = env.pop(1)?;
                 let cond = env.pop(2)?;